requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
python-dotenv
WTForms-SQLAlchemy
